requirements.sink-source.txt
# requirements.sink.txt
kafka-python==2.0.2
mysql-connector-python==8.0.27
cassandra-driver==3.25.0